% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{assert_bbox_is_valid_size}
\alias{assert_bbox_is_valid_size}
\title{Assert Bbox is a Valid Size}
\usage{
assert_bbox_is_valid_size(bbox, min_area = 10000, max_area = 1e+09)
}
\arguments{
\item{bbox}{A \code{SpatExtent} defining the bbox.}

\item{min_area}{The minimum allowable area in meters.}

\item{max_area}{The maximum allowable area in meters.}
}
\description{
Raise an error if the given bounding box has an area smaller than
\code{min_area} or larger than \code{max_area}.
}
