% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreation.R
\name{compute_component}
\alias{compute_component}
\title{Compute Component}
\usage{
compute_component(component, persona, data_dir, bbox = NULL)
}
\arguments{
\item{component}{The name of the component to compute (one of "SLSRA", "FIPS_N", "FIPS_I", "Water").}

\item{persona}{A named vector containing the persona scores (for all components).}

\item{data_dir}{Path to the directory containing the rasters.}

\item{bbox}{An optional bounding box for cropping.}
}
\value{
A single-layered \code{SpatRaster} which is the contribution to the Recreational Potential from this component.
}
\description{
Compute a single component of the Recreational Potential.
}
\details{
This function computes one \emph{component}, i.e. one of (i) SLSRA, (ii) FIPS_N,
(iii) FIPS_I and (iv) Water, of the Recreational Potential. This is simply
a pixel-wise summation of all of the layers (most of which are either 1/present
or 0/absent, weighted by the corresponding scores.

\deqn{ C(x, y) = \sum_{i=1}^{n_\text{feat}} s_i z_i(x, y) }
}
