% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R, R/potential.R
\name{compute_potential}
\alias{compute_potential}
\title{Compute Recreational Potential}
\usage{
compute_potential(persona, raster_dir, bbox = NULL)

compute_potential(persona, raster_dir, bbox = NULL)
}
\description{
Compute Recreational Potential

Compute Recreational Potential
}
