% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint_sigma2.R
\name{confint_sigma2}
\alias{confint_sigma2}
\title{Confidence interval for sigma2 in lm objects.}
\usage{
confint_sigma2(object, level = 0.95)
}
\arguments{
\item{object}{a \link{lm} object.}

\item{level}{the confidence level required.}
}
\value{
A matrix with columns giving lower and upper confidence limits.
}
\description{
The \code{confint_sigma2} function computes confidence interval for the sigma2 parameter in a linear model fitted with the \link{lm} function.
}
\examples{
mod <- lm(dist ~ speed, data=cars)
confint_sigma2(object=mod, level=0.95)
confint_sigma2(object=mod, level=0.80)
}
