% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_param.R
\name{est_param}
\alias{est_param}
\title{Estimation of mean and/or variance for models fitted with gamlss.}
\usage{
est_param(mod, fun = "mean", m = 100)
}
\arguments{
\item{mod}{a gamlss object.}

\item{fun}{the estimated parameter, by default is mean but can be change for sd or var.}

\item{m}{a number to indicate the number of observations to simulate, by default its values is 100.}
}
\value{
\code{est_param} function returns a vector.
}
\description{
This function estimates parameters like mean, standard deviation for a gamlss object. This function is very useful when the mean or standard deviation do not match with mu, sigma or some function of mu and/or sigma.
}
\details{
The function obtains the fitted values for mu, sigma, nu and tau. The functions simulates m observations and then it obtains the mean/fun defined by the user.
}
\examples{

# Example for NOrmal regression ------------------------
# Here E(y) = mu and Var(y)= sigma
n <- 7
x <- runif(n=n)
y <- rnorm(n=n, mean=-5 + 10 * x, sd=exp(1 + 2 * x))
library(gamlss)
mod <- gamlss(y ~ x, sigma.fo=~x, family=NO)

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- est_param(mod, m=10000, fun="mean")
cbind(m1, m2) # Comparing the estimated means
cor(m1, m2)

# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma")
s2 <- est_param(mod, m=10000, fun="sd")
cbind(s1, s2) # Comparing the estimated standard deviations
cor(s1, s2)

# Example for GAmma regression --------------------------
# Here E(y) = mu and Var(y)= sigma^2 mu^2
n <- 7
x <- runif(n=n)
y <- rGA(n=n, mu=exp(-5 + 10 * x), sigma=exp(-1 + 2 * x))
mod <- gamlss(y ~ x, sigma.fo=~x, family=GA)

# Obtaining the means
m1 <- fitted(mod, what="mu")
m2 <- est_param(mod, m=10000, fun="mean")
cbind(m1, m2) # Comparing the estimated means
cor(m1, m2)

# Obtaining the standard deviations
s1 <- fitted(mod, what="sigma") * fitted(mod, what="mu")
s2 <- est_param(mod, m=10000, fun="sd")
cbind(s1, s2) # Comparing the estimated standard deviations
cor(s1, s2)

}
