% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_model_samp.R
\name{full_model_samp}
\alias{full_model_samp}
\title{Full version of the model}
\usage{
full_model_samp(sig_gl, sig_rel, n_targ, n_dist, sim, t_type = FALSE,
  opt = FALSE)
}
\arguments{
\item{sig_gl}{Signal strength for detecting position change for a single sphere}

\item{sig_rel}{Signal strength for detecting relative position change for a single sphere}

\item{n_targ}{The number of targets, i.e. spheres that move on that trial}

\item{n_dist}{The number of static distractors (unless t_type=TRUE), in which case it is the number of spheres moving with lesser signal strength}

\item{sim}{The number of trials to simulate}

\item{t_type}{The type of the target/distractor relationship}

\item{opt}{An existing d' to compare predictions to if given}
}
\description{
Returns either the d-prime of a condition given by the inputs or the absolute error of the predicted d' relative to a prespecified value.
}
\examples{
full_model_samp()
}
\keyword{fullmodel}
