% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_score_mappings}
\alias{get_score_mappings}
\title{Get Score Mappings}
\usage{
get_score_mappings(config, persona)
}
\arguments{
\item{config}{\code{data.frame} The configuration file}

\item{persona}{\code{integer} The persona}
}
\value{
A mapping {layer_name : {raster_value : persona_score}}
}
\description{
Get Score Mappings
}
