% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_config}
\alias{load_config}
\title{Load configuration from file}
\usage{
load_config(config_path = NULL)
}
\arguments{
\item{config_path}{\code{character} Optional path to non-default configuration file}
}
\description{
Load configuration from file
}
