% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{load_raster}
\alias{load_raster}
\title{Load a cropped raster}
\usage{
load_raster(raster_path, area = NULL)
}
\arguments{
\item{raster_path}{\code{character} Path to a raster file to load}

\item{crop_area}{A SpatExtent or another object with a SpatExtent}
}
\description{
Load a cropped raster from a file
}
\details{
Details...
}
\examples{
load_raster("path/to/raster.tif", terra::vect("path/to/shapefile.shp"))
load_raster("path/to.raster.tif", terra::ext(xmin, xmax, ymin, ymax))

}
