% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logistic_func}
\alias{logistic_func}
\title{Compute a logistic function}
\usage{
logistic_func(x, alpha, kappa)
}
\arguments{
\item{x}{A raster}

\item{alpha}{Coefficient in the exponent}

\item{kappa}{A less important parameter}
}
\description{
Compute a logistic function
}
