% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{map_distance_to_unit_interval}
\alias{map_distance_to_unit_interval}
\title{Map distances to the unit interval}
\usage{
map_distance_to_unit_interval(x, alpha, kappa)
}
\arguments{
\item{x}{A raster}

\item{alpha}{Coefficient in the exponent}

\item{kappa}{A less important parameter}
}
\description{
Uses a logistic function to map positive distances \eqn{d}
to the unit interval \eqn{x \in [0, 1]}.
}
\details{
\deqn{ x = \frac{\kappa + 1}{\kappa + \exp(\alpha d)} }
}
