% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_production.R, R/data.R
\name{na_to_zero}
\alias{na_to_zero}
\title{NA to zero}
\usage{
na_to_zero(raster)

na_to_zero(raster)
}
\description{
Map NA (not available / missing) to zero, keeping NaN as is

Map NA (not available / missing) to zero, keeping NaN as is
}
