% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obsfa.R
\name{plot_obsfa}
\alias{plot_obsfa}
\title{Plot observations in the coordinate plane Fi-Fj.}
\usage{
plot_obsfa(scores, fx = 1, fy = 2, lchar = 3, ...)
}
\arguments{
\item{scores}{a matrix obtained from \code{\link{factanal}} function.}

\item{fx}{a number to indicate the factor to plot in x axis.}

\item{fy}{a number to indicate the factor to plot in y axis.}

\item{lchar}{the number of first characters to identify points, by default is 3.}

\item{...}{arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
\code{plot_obsfa} function returns a plot with the 
observations in the coordinate plane Fi-Fj from a factor analysis.
}
\description{
This function plots the original observations in the coordinate plane Fi-Fj from a factor analysis.
}
\details{
The number in \code{fx} and \code{fy} must be less or equal
number of factors in \code{factanal} function.
}
\examples{
v1 <- c(1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,4,5,6)
v2 <- c(1,2,1,1,1,1,2,1,2,1,3,4,3,3,3,4,6,5)
v3 <- c(3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,5,4,6)
v4 <- c(3,3,4,3,3,1,1,2,1,1,1,1,2,1,1,5,6,4)
v5 <- c(1,1,1,1,1,3,3,3,3,3,1,1,1,1,1,6,4,5)
v6 <- c(1,1,1,2,1,3,3,3,4,3,1,1,1,2,1,6,5,4)
m1 <- cbind(v1,v2,v3,v4,v5,v6)

scores <- factanal(m1, factors=3, method="mle",
                   scores="Bartlett")$scores

plot_obsfa(scores, fx=1, fy=2, col='blue', pch=20)
plot_obsfa(scores, fx=1, fy=3, col='red', pch=19)
}
