% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obspc.R
\name{plot_obspc}
\alias{plot_obspc}
\title{Plot observations in the coordinate plane PC1-PC2.}
\usage{
plot_obspc(mod, lchar = 3, pch = 20, ...)
}
\arguments{
\item{mod}{a prcomp or princomp object obtained from \code{\link{prcomp}} or \code{\link{princomp}} functions.}

\item{lchar}{the number of first characters to identify points.}

\item{pch}{an integer specifying a symbol or a single character to be used as the default in plotting points.}

\item{...}{arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
The function returns a list with two elements, coordinate points for first
and second dimensions.
}
\description{
This function plots the observations in the coordinate plane PC1-PC2.
}
\examples{
# For a prcomp object
pca1 <- prcomp(USArrests, scale=TRUE)  # Example
plot_obspc(pca1)  
# Default plot
plot_obspc(pca1, lchar=3, col='blue', pch=8)  # Customized plot
#
# For a princomp object
pca2 <- princomp(x=USArrests, cor=TRUE)
plot_obspc(pca2, col='red', ylim=c(-4, 4), xlim=c(-4, 4)) 
}
