% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prof.R
\name{plot_prof}
\alias{plot_prof}
\title{Profile plot}
\usage{
plot_prof(x, Row = TRUE, cex.names = 0.5, xlab = "",
  ylab = "Frequency", cex = 1, ...)
}
\arguments{
\item{x}{a table, xtabs or matrix object.}

\item{Row}{a logical value to indicate if profile plot is for Rows o Columns.}

\item{cex.names}{Expansion factor for axis names (bar labels).}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis, by default is Frequency.}

\item{cex}{a character expansion factor for legend.}

\item{...}{arguments to be passed to methods, such as graphical parameters (see par).}
}
\value{
\code{plot_prof} function returns a profile plot.
}
\description{
This function gives the profile plot for contingency table objects created by table or xtabs or matrix.
}
\examples{
\dontrun{
require(MASS)
# Using a table object
tbl1 <- with(Cars93, table(Type, AirBags))
plot_prof(x=tbl1, Row=TRUE)
plot_prof(x=tbl1, Row=FALSE, cex.names=0.9, cex=0.6)
#---------------------
# Using a xtabs object
tbl2 <- with(Cars93, xtabs(~ Type + AirBags))
plot_prof(x=tbl2, Row=TRUE, cex.names=1.8)
}
}
