% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2predict.R
\name{r2predict}
\alias{r2predict}
\title{Calculates R2 prediction statistic.}
\usage{
r2predict(mod)
}
\arguments{
\item{mod}{A lm object.}
}
\value{
The \code{press} function returns the press statistic defined as \eqn{R2predict=1-PRESS/SST}.
}
\description{
This function calculates R2 prediction statistic.
}
\examples{
x <- c(1, 1, 2, 3, 3, 4, 4, 4, 4.7, 5, 5.6, 5.6, 5.6, 6, 6, 6.5, 6.9)
y <- c(10.84, 9.3, 16.35, 22.88, 24.35, 24.56, 25.86, 29.16, 24.59,
       22.25, 25.9, 27.2, 25.61, 25.45, 26.56, 21.03, 21.46)
mod <- lm(y ~ x)
r2predict(mod)
}
