% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reclassify}
\alias{reclassify}
\title{Reclassify a SpatRaster based on a lookup table}
\usage{
reclassify(spat_raster, mapping)
}
\arguments{
\item{spat_raster}{\code{SpatRaster} The SpatRaster to be reflassified}

\item{mapping}{\code{numeric} A \emph{named} vector mapping current values (names) to new (values)}
}
\description{
Reclassify a SpatRaster based on a lookup table
}
