% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{reclassify_layer}
\alias{reclassify_layer}
\title{Reclassify a single-layered SpatRaster using a mapping}
\usage{
reclassify_layer(layer, mapping)
}
\arguments{
\item{mapping}{\code{numeric} A mapping from current values to new values}

\item{spat_raster}{\code{SpatRaster} The SpatRaster to be reflassified}
}
\description{
Reclassify a single-layered SpatRaster using a mapping
}
