% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_model.R
\name{rel_model}
\alias{rel_model}
\title{Relative-only version of the model}
\usage{
rel_model(sig_rel, n_targ, n_dist, sim, t_type = FALSE, opt = FALSE)
}
\arguments{
\item{sig_rel}{Signal strength for detecting relative position change for a single sphere}

\item{n_targ}{The number of targets, i.e. spheres that move on that trial}

\item{n_dist}{The number of static distractors (unless t_type=TRUE), in which case it is the number of spheres moving with lesser signal strength}

\item{sim}{The number of trials to simulate}

\item{t_type}{The type of the target/distractor relationship}

\item{opt}{An existing d' to compare predictions to if given}
}
\description{
Returns either the d-prime of a condition given by the inputs or the absolute error of the predicted d' relative to a prespecified value.
}
\examples{
rel_model()
}
\keyword{fullmodel}
