% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res.R
\name{Res}
\alias{Res}
\title{Compute residuals for \code{lm} objects.}
\usage{
Res(mod)
}
\arguments{
\item{mod}{A lm object.}
}
\value{
\code{Res} function returns a dataframe with residuals.
}
\description{
\code{Res} function computes Standardized Residuals,
Studentized Residuals, Press Residuals or Deleted Residuals
and Externally studentized residuals.
}
\examples{
# Multiple linear regression
mod <- lm(mpg ~ hp + wt, data=mtcars)
#
res <- Res(mod)  # Residuals
res$ri  # To obtain Studentized Residuals (ri)
#
par(mfrow=c(2, 2))
plot(res$di, ylab=expression(d[i]))
abline(h=-3, lty='dotted')
abline(h=+3, lty='dotted')
plot(res$ri, ylab=expression(r[i]))
plot(res$pres, ylab=expression(e[(i)]))
plot(res$ti, ylab=expression(t[i]))
}
