% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{save_persona}
\alias{save_persona}
\title{Save Persona}
\usage{
save_persona(persona, csv_path, name)
}
\arguments{
\item{persona}{\code{integer} A vector of integers representing the persona}

\item{csv_path}{\code{character} Path to write a csv file, which may already exist}

\item{name}{\code{character} Name of the persona}
}
\description{
Saves a persona to a csv file. If the csv file already exists, the persona
will be appended as a new column, unless the provided name matches an existing
column, in which case it will overwrite the existing data.
}
