% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{suggestNextWords}
\alias{suggestNextWords}
\title{Next word predictor}
\usage{
suggestNextWords(phrase, ngrams, n = NUM_SUGGESTIONS)
}
\arguments{
\item{phrase}{A character vector representing the preceeding words for the prediction}

\item{ngrams}{A vector of `quanteda` N-grams, in ascending order (e.g. 1-gram,
2-gram, etc.)}

\item{n}{The amount of predicted words.}
}
\value{
A vector of predicted words, ordered by descending likelihood.
}
\description{
This function generates `n` word suggestions (3 by
default) using a list of `quanteda` N-grams as observed data, and
a phrase for which you want the prediction to be based of.
}
\details{
It implements a simplified version of the [Katz's backoff model]
(https://en.wikipedia.org/wiki/Katz%27s_back-off_model).
}
