% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_data_production.R, R/data.R
\name{sum_layers}
\alias{sum_layers}
\title{Sum the layers of a SpatRaster}
\usage{
sum_layers(raster)

sum_layers(raster)
}
\description{
Sum the layers of a SpatRaster

Sum the layers of a SpatRaster
}
