% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_lmer.R
\name{summary_lmer}
\alias{summary_lmer}
\title{Summary table for a \code{lmerMod} object.}
\usage{
summary_lmer(mod)
}
\arguments{
\item{mod}{A `lmerMod` object.}
}
\value{
\code{summary_lmer} function returns a matrix with p values in last column.
}
\description{
This function creates the summary table with p values for a \code{lmerMod} object.
}
\examples{
\dontrun{
require(lme4)
fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
summary(fm1) # Usual table without p values
summary_lmer(fm1) # Table with p values
}
}
