% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute_potential}
\alias{compute_potential}
\title{Compute Recreational Potential}
\usage{
compute_potential(crop_area, persona, data_dir, config_path = NULL)
}
\description{
Compute Recreational Potential
}
