% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load Config}
\usage{
load_config(config_path = NULL)
}
\arguments{
\item{config_path}{(\code{character}) Path to a CSV file containing the configuration.
If no path is given, the default configuration from \code{inst/extdata/config.csv} will be loaded.}
}
\value{
A \code{data.frame} containing the configuration.
}
\description{
Load a model configuration from a CSV file.
}
\details{
This function essentially calls \code{readr::read_csv} with \code{col_types} set to reflect the
expected columns in a model configuration. That is: four \code{character} columns for
\code{Component}, \code{Dataset}, \code{Name}, and \code{Description}, followed by one \code{integer} column
for \code{Raster_Val}. See \code{inst/extdata/config.csv} for an example of a valid configuration
file.
}
